More refactoring, managed to get pip package working properly
