+++ kbf
@@ -1,0 +1,5 @@
+#!/usr/bin/python
+
+# This file provides a shorter executable instead of calling "python kuberfest/kuberfest2.py" each time
+
+from kuberfest import run
