+++ kuberfest.egg-info/SOURCES.txt
@@ -1,7 +1,29 @@
 README.md
+kbf
 setup.py
-bin/kbf
+dotnet_demo_project/kuberfest/functions.py
+dotnet_demo_project/kuberfest/settings.py
+dotnet_demo_project/kuberfest/variables.py
+kuberfest/consts.py
+kuberfest/project.py
+kuberfest/run.py
 kuberfest.egg-info/PKG-INFO
 kuberfest.egg-info/SOURCES.txt
 kuberfest.egg-info/dependency_links.txt
 kuberfest.egg-info/top_level.txt
+kuberfest/commands/__init__.py
+kuberfest/commands/build.py
+kuberfest/commands/context.py
+kuberfest/commands/delete.py
+kuberfest/commands/deploy.py
+kuberfest/commands/development.py
+kuberfest/commands/init_db.py
+kuberfest/commands/minikube_ip.py
+kuberfest/commands/print_environment.py
+kuberfest/commands/push.py
+kuberfest/commands/start_minikube.py
+kuberfest/tools/__init__.py
+kuberfest/tools/base.py
+kuberfest/tools/debug.py
+kuberfest/tools/kubernetes.py
+kuberfest/tools/yaml.py
