+++ kuberfest.egg-info/top_level.txt
@@ -1,1 +1,4 @@
-
+build
+dist
+dotnet_demo_project
+kuberfest
