+++ kuberfest/commands/__init__.py
@@ -1,7 +1,7 @@
 import importlib
 import argparse
-import consts
-from tools.debug import Debug
+from kuberfest import consts
+from kuberfest.tools.debug import Debug
 
 
 # Commands will be run in the same order of the dictionary
