+++ kuberfest/commands/build.py
@@ -1,6 +1,5 @@
-import tools
 import os
-from tools.debug import Debug
+from kuberfest.tools.debug import Debug
 
 
 def run(project, value):
