+++ kuberfest/commands/development.py
@@ -1,5 +1,3 @@
-import tools
-
 def run(project, value):
     project.set_development(value)
